requirements.txt
# black
pre-commit
mypy
flake8
tensorflow-gpu>=1.8
